# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/journals:
    get:
      tags:
        - Journals
      summary: List Journals
      description: Get a paginated list of Journals with optional filtering and sorting
      operationId: listJournals
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Journals
      summary: Create Journal
      description: Create a new Journal
      operationId: createJournal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Journal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/journal_lines:
    get:
      tags:
        - JournalLines
      summary: List JournalLines
      description: Get a paginated list of JournalLines with optional filtering and sorting
      operationId: listJournalLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - JournalLines
      summary: Create JournalLine
      description: Create a new JournalLine
      operationId: createJournalLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Journal:
      type: object
      required:
        - id
        - company_id
        - journal_number
        - journal_type
        - transaction_date
        - description
        - currency
        - total_debit
        - total_credit
        - line_count
        - source
        - is_reversed
        - is_reversing
        - auto_reverse
        - status
        - requires_approval
        - is_voided
        - attachments
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        journal_number:
          type: string
          maxLength: 30
        reference_number:
          nullable: true
          type: string
          maxLength: 50
        journal_type:
          type: string
        branch_id:
          nullable: true
          type: string
          format: uuid
        transaction_date:
          type: string
          format: date
        posting_date:
          nullable: true
          type: string
          format: date
        fiscal_period_id:
          nullable: true
          type: string
          format: uuid
        fiscal_year:
          nullable: true
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        fiscal_month:
          nullable: true
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        description:
          type: string
          maxLength: 500
        notes:
          nullable: true
          type: string
          maxLength: 1000
        currency:
          type: string
        total_debit:
          type: string
        total_credit:
          type: string
        line_count:
          type: integer
          format: int32
        source:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        is_reversed:
          type: boolean
        reversed_by_id:
          nullable: true
          type: string
          format: uuid
        reversed_at:
          nullable: true
          type: string
          format: date-time
        reversal_reason:
          nullable: true
          type: string
          maxLength: 255
        reverses_id:
          nullable: true
          type: string
          format: uuid
        is_reversing:
          type: boolean
        auto_reverse:
          type: boolean
        auto_reverse_date:
          nullable: true
          type: string
          format: date
        status:
          type: string
        requires_approval:
          type: boolean
        approval_threshold:
          nullable: true
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        submitted_by:
          nullable: true
          type: string
          format: uuid
        approved_at:
          nullable: true
          type: string
          format: date-time
        approved_by:
          nullable: true
          type: string
          format: uuid
        rejected_at:
          nullable: true
          type: string
          format: date-time
        rejected_by:
          nullable: true
          type: string
          format: uuid
        rejection_reason:
          nullable: true
          type: string
          maxLength: 500
        posted_at:
          nullable: true
          type: string
          format: date-time
        posted_by:
          nullable: true
          type: string
          format: uuid
        is_voided:
          type: boolean
        voided_at:
          nullable: true
          type: string
          format: date-time
        voided_by:
          nullable: true
          type: string
          format: uuid
        void_reason:
          nullable: true
          type: string
          maxLength: 500
        attachments:
          type: object
        metadata:
          type: object
...
    CreateJournalRequest:
      type: object
      required:
        - company_id
        - journal_number
        - journal_type
        - transaction_date
        - description
        - currency
        - total_debit
        - total_credit
        - line_count
        - source
        - is_reversed
        - is_reversing
        - auto_reverse
        - status
        - requires_approval
        - is_voided
        - attachments
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        journal_number:
          type: string
          maxLength: 30
        reference_number:
          nullable: true
          type: string
          maxLength: 50
        journal_type:
          type: string
        branch_id:
          nullable: true
          type: string
          format: uuid
        transaction_date:
          type: string
          format: date
        posting_date:
          nullable: true
          type: string
          format: date
        fiscal_period_id:
          nullable: true
          type: string
          format: uuid
        fiscal_year:
          nullable: true
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        fiscal_month:
          nullable: true
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        description:
          type: string
          maxLength: 500
        notes:
          nullable: true
          type: string
          maxLength: 1000
        currency:
          type: string
        total_debit:
          type: string
        total_credit:
          type: string
        line_count:
          type: integer
          format: int32
        source:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        is_reversed:
          type: boolean
        reversed_by_id:
          nullable: true
          type: string
          format: uuid
        reversed_at:
          nullable: true
          type: string
          format: date-time
        reversal_reason:
          nullable: true
          type: string
          maxLength: 255
        reverses_id:
          nullable: true
          type: string
          format: uuid
        is_reversing:
          type: boolean
        auto_reverse:
          type: boolean
        auto_reverse_date:
          nullable: true
          type: string
          format: date
        status:
          type: string
        requires_approval:
          type: boolean
        approval_threshold:
          nullable: true
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        submitted_by:
          nullable: true
          type: string
          format: uuid
        approved_at:
          nullable: true
          type: string
          format: date-time
        approved_by:
          nullable: true
          type: string
          format: uuid
        rejected_at:
          nullable: true
          type: string
          format: date-time
        rejected_by:
          nullable: true
          type: string
          format: uuid
        rejection_reason:
          nullable: true
          type: string
          maxLength: 500
        posted_at:
          nullable: true
          type: string
          format: date-time
        posted_by:
          nullable: true
          type: string
          format: uuid
        is_voided:
          type: boolean
        voided_at:
          nullable: true
          type: string
          format: date-time
        voided_by:
          nullable: true
          type: string
          format: uuid
        void_reason:
          nullable: true
          type: string
          maxLength: 500
        attachments:
          type: object
        metadata:
          type: object
...
    JournalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Journal'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    JournalLine:
      type: object
      required:
        - id
        - journal_id
        - company_id
        - line_number
        - account_id
        - account_number
        - account_name
        - debit_amount
        - credit_amount
        - currency
        - exchange_rate
        - base_debit_amount
        - base_credit_amount
        - is_tax_line
        - has_quantity
        - is_reconciled
        - is_posted
        - tags
        - data
      properties:
        id:
          type: string
          format: uuid
        journal_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        line_number:
          type: integer
          format: int32
        account_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_name:
          type: string
          maxLength: 100
        debit_amount:
          type: string
        credit_amount:
          type: string
        currency:
          type: string
        exchange_rate:
          type: string
        base_debit_amount:
          type: string
        base_credit_amount:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 255
        cost_center_id:
          nullable: true
          type: string
          format: uuid
        project_id:
          nullable: true
          type: string
          format: uuid
        department_id:
          nullable: true
          type: string
          format: uuid
        dimensions:
          nullable: true
          type: object
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        is_tax_line:
          type: boolean
        tax_rate:
          nullable: true
          type: string
        tax_base_amount:
          nullable: true
          type: string
        related_line_id:
          nullable: true
          type: string
          format: uuid
        has_quantity:
          type: boolean
        quantity:
          nullable: true
          type: string
        unit:
          nullable: true
          type: string
          maxLength: 20
        unit_price:
          nullable: true
          type: string
        is_reconciled:
          type: boolean
        reconciliation_id:
          nullable: true
          type: string
          format: uuid
        reconciled_at:
          nullable: true
          type: string
          format: date-time
        full_reconcile_id:
          nullable: true
          type: string
          format: uuid
        is_posted:
          type: boolean
        ledger_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        tags:
          type: object
        data:
          type: object
...
    CreateJournalLineRequest:
      type: object
      required:
        - journal_id
        - company_id
        - line_number
        - account_id
        - account_number
        - account_name
        - debit_amount
        - credit_amount
        - currency
        - exchange_rate
        - base_debit_amount
        - base_credit_amount
        - is_tax_line
        - has_quantity
        - is_reconciled
        - is_posted
        - tags
        - data
      properties:
        journal_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        line_number:
          type: integer
          format: int32
        account_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_name:
          type: string
          maxLength: 100
        debit_amount:
          type: string
        credit_amount:
          type: string
        currency:
          type: string
        exchange_rate:
          type: string
        base_debit_amount:
          type: string
        base_credit_amount:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 255
        cost_center_id:
          nullable: true
          type: string
          format: uuid
        project_id:
          nullable: true
          type: string
          format: uuid
        department_id:
          nullable: true
          type: string
          format: uuid
        dimensions:
          nullable: true
          type: object
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        is_tax_line:
          type: boolean
        tax_rate:
          nullable: true
          type: string
        tax_base_amount:
          nullable: true
          type: string
        related_line_id:
          nullable: true
          type: string
          format: uuid
        has_quantity:
          type: boolean
        quantity:
          nullable: true
          type: string
        unit:
          nullable: true
          type: string
          maxLength: 20
        unit_price:
          nullable: true
          type: string
        is_reconciled:
          type: boolean
        reconciliation_id:
          nullable: true
          type: string
          format: uuid
        reconciled_at:
          nullable: true
          type: string
          format: date-time
        full_reconcile_id:
          nullable: true
          type: string
          format: uuid
        is_posted:
          type: boolean
        ledger_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        tags:
          type: object
        data:
          type: object
...
    JournalLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JournalLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []